backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-environ==0.11.2
django-redis==5.4.0
redis==5.2.1
Pillow==11.1.0
cos-python-sdk-v5==1.9.38
requests==2.32.3
PyMySQL==1.1.1
gunicorn==23.0.0
